requirements.txt
playwright
pyyaml
openai
pillow
